backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
anthropic==0.34.2
apscheduler==3.10.4
python-multipart==0.0.12
aiofiles==24.1.0
pillow==10.4.0
python-magic==0.4.27
pytz==2024.2
tenacity==9.0.0
python-dateutil==2.9.0.post0
